+++ include/solarus/lua/LuaContext.h
@@ -580,6 +580,8 @@ class LuaContext {
       main_api_get_metatable,
       main_api_get_os,
       main_api_get_game,
+      main_api_rawget,
+      main_api_rawset,
 
       // Audio API.
       audio_api_get_sound_volume,
@@ -1294,6 +1296,7 @@ class LuaContext {
       state_api_set_carried_object_action,
 
       // available to all userdata types
+      userdata_rawget_as_table,
       userdata_meta_gc,
       userdata_meta_newindex_as_table,
       userdata_meta_index_as_table,