+++ tests/cmake/AddTestMaps.cmake
@@ -151,6 +151,7 @@ list(APPEND LUA_TEST_MAPS
   "bugs/1416_state_set_affected_by_ground_speed"
   "bugs/1418_state_set_can_be_hurt"
   "bugs/1423_crash_carried_without_shadow"
+  "bugs/1424_rawget_rawset"
   "bugs/1428_movement_blocking"
   "bugs/1436_thrown_object_shadow"
   "bugs/1439_ground_sprite_ladder"
@@ -159,6 +160,8 @@ list(APPEND LUA_TEST_MAPS
   "bugs/1452_crash_carried_object"
   "bugs/1458_ground_sprite_when_jumping"
   "bugs/1461_tileset_sprite"
+  "bugs/1471_pickable_wrong_layer"
+  "bugs/1476_crash_command_binding"
 )
 
 # Test that should not fail when there are errors